Restructure components by folders
