+++ canteen-management-system-client/src/components/LandingPage/LandingPageRight.jsx
