+++ canteen-management-system-client/src/components/LandingPage/LogInForm.jsx
@@ -7,7 +7,7 @@ import {
 import { Button, Col, Divider, Input, Row, Typography } from "antd";
 import "./LogInForm.css";
 import { useRef, useState, useEffect, useContext } from "react";
-import AuthContext from "../context/AuthProvider";
+import AuthContext from "../../context/AuthProvider";
 import { useNavigate } from "react-router-dom";
